requirements.txt
backcall==0.1.0
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
debug==0.3.2
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.2
django-appconf==1.0.3
django-autoslug==1.9.4
django-compressor==2.3
django-constance==2.4.0
django-cors-headers==2.4.0
django-debug-toolbar==1.11
django-extensions==2.1.7
django-filter==2.1.0
django-grappelli==2.12.3
django-jet==1.0.8
django-mail-panel==1.2.2
django-picklefield==2.0
django-post-office==3.2.1
django-templated-mail==1.1.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
django-mptt==0.10.0
djangorestframework-recursive==0.1.2
djoser==1.7.0
drf-nested-routers==0.90.0
easy-thumbnails==2.6
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.13.3
jsonfield==2.0.2
lxml==4.3.3
odfpy==1.4.0
openpyxl==2.6.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
PyJWT==1.7.1
pytils==0.3
pytz==2019.1
PyYAML==5.1
raven==6.10.0
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
see==1.4.1
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.4
xlrd==1.2.0
xlwt==1.3.0
django-mptt-admin==0.7.2
celery==4.3.0
selenium==3.141.0

